#5424: remove metal_ckernel_spfu ops for grayskull
