+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/ckernel_sfpu_comp.h
@@ -1,0 +1,92 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel.h"
+#include "ckernel_defs.h"
+#include "noc_nonblocking_api.h"
+
+using namespace sfpi;
+
+namespace ckernel {
+namespace sfpu {
+
+template <bool APPROXIMATION_MODE, SfpuType COMP_MODE, int ITERATIONS = 4>
+inline void calculate_comp(uint exponent_size_8)
+{
+    const vFloat zero = 0.0f;
+    const vFloat one = 1.0f;
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat v = dst_reg[0];
+        vFloat flag1, flag2;
+
+	//a[i] == 0
+	if constexpr(COMP_MODE == SfpuType::equal_zero) {
+	    v_if (_sfpu_is_fp16_zero_(v, exponent_size_8)) {
+	        v = one;
+	    } v_else {
+	        v = zero;
+	    }
+	    v_endif;
+	  }
+
+	//a[i] != 0
+	if constexpr(COMP_MODE == SfpuType::not_equal_zero) {
+	    v_if (_sfpu_is_fp16_zero_(v, exponent_size_8)) {
+	        v = zero;
+	    } v_else {
+	        v = one;
+	    }
+	    v_endif;
+        }
+
+	//a[i] < 0
+	if constexpr(COMP_MODE == SfpuType::less_than_zero) {
+	    v_if (v >= 0.0f) {
+	        v = zero;
+	    } v_else {
+	        v = one;
+	    }
+	    v_endif;
+        }
+
+	//a[i] >= 0
+	if constexpr(COMP_MODE == SfpuType::greater_than_equal_zero) {
+	    v_if (v >= 0.0f) {
+	        v = one;
+	    } v_else {
+	        v = zero;
+	    }
+	    v_endif;
+        }
+
+	//a[i] > 0
+	if constexpr(COMP_MODE == SfpuType::greater_than_zero) {
+	    v_if (v > 0.0f) {
+	        v = one;
+	    } v_else {
+	        v = zero;
+	    }
+	    v_endif;
+        }
+
+	//a[i] <= 0
+	if constexpr(COMP_MODE == SfpuType::less_than_equal_zero) {
+	    v_if (v > 0.0f) {
+	        v = zero;
+	    } v_else {
+	        v = one;
+	    }
+	    v_endif;
+        }
+
+	dst_reg[0] = v;
+	dst_reg++;
+    }
+}
+
+}  // namespace sfpu
+}  // namespace ckernel
