+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/ckernel_sfpu_exp2.h
@@ -14,34 +14,29 @@ using namespace sfpi;
 namespace ckernel {
 namespace sfpu {
 
-template <bool APPROXIMATION_MODE, int ITERATIONS = 8>
+template <bool APPROXIMATION_MODE, int ITERATIONS = 4>
 inline void calculate_exp2()
 {
+    constexpr bool zero_negative = true;
     // SFPU microcode
     for (int d = 0; d < ITERATIONS; d++)
     {
         vFloat v = dst_reg[0];
+        // y = exp(x * log(2))
         // log(2) = 0.6931471805;
         v = v * 0.6931471805f;
-	    // exp = e^(v)
-	    vFloat exp = calculate_exponential_body_improved<APPROXIMATION_MODE, true>(v);
-	    dst_reg[0] = exp;
+        vFloat exp = calculate_exponential_body_improved<APPROXIMATION_MODE, zero_negative>(v);
+        dst_reg[0] = exp;
         dst_reg++;
     }
 }
 
 template <bool APPROXIMATION_MODE>
-inline void exp2_init() {
-
-    if constexpr (APPROXIMATION_MODE) {
-        vConstFloatPrgm0 = 1.442695f; // ln2_recip
-        vConstFloatPrgm1 = s2vFloat16b(p_exp::C23_73);
-        vConstFloatPrgm2 = s2vFloat16b(p_exp::ADJ_EXP);
-    }
-    else{
-        vConstFloatPrgm0 = 1.442695f; // ln2_recip
-        vConstFloatPrgm1 = 2.0f;
-        vConstFloatPrgm2 = 0.863281f;
+inline void exp2_init()
+{
+    if constexpr(APPROXIMATION_MODE) {
+        TTI_SFPLOADI(p_sfpu::LREG0, 0, p_exp::C23_73);
+        TTI_SFPLOADI(p_sfpu::LREG2, 0, p_exp::ADJ_EXP);
     }
 }
