+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/ckernel_sfpu_power_iterative.h
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel.h"
+#include "ckernel_defs.h"
+#include "noc_nonblocking_api.h"
+
+using namespace sfpi;
+
+namespace ckernel {
+namespace sfpu {
+
+template <bool APPROXIMATION_MODE,int ITERATIONS = 4>
+inline void calculate_power_iterative(const uint exponent)
+{
+    #pragma GCC unroll 4
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat in = dst_reg[0];
+        vFloat result = 1.0f;
+        for (uint i = 0; i < exponent; i++) {
+            result *= in;
+        }
+	    dst_reg[0]=result;
+        dst_reg++;
+    }
+}
+
+}  // namespace sfpu
+}  // namespace ckernel
