+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/ckernel_sfpu_trigonometry.h
@@ -8,6 +8,7 @@
 #include "ckernel_defs.h"
 #include "noc_nonblocking_api.h"
 #include "sfpi.h"
+#include "ckernel_sfpu_recip.h"
 
 using namespace sfpi;
 
@@ -211,19 +212,113 @@ inline void calculate_sfpu_trig() {
     }
 }
 
+#define POLYVAL6(coef5, coef4, coef3, coef2, coef1, coef0, t4)  (t4 * (t4 * (t4 * (t4 * (coef5 * t4 + coef4) + coef3) + coef2) + coef1) + coef0)
+
 template <bool APPROXIMATION_MODE>
-void sine_init() {
-    ;
+sfpi_inline vFloat sfpu_atan_maclaurin_series(vFloat val)
+{
+    v_if(1 > sfpi::abs(val)){
+        dst_reg[0] = sfpi::abs(val)  ;
+    }
+    v_else{
+        dst_reg[0] =  sfpu_reciprocal<true>(sfpi::abs(val));
+    }
+    v_endif;
+
+    vFloat t1 = dst_reg[0] * dst_reg[0];
+
+    t1 = POLYVAL6(-0.013480470f, 0.057477314f, -0.121239071f, 0.195635925f, -0.332994597f, 0.999995630f, t1);
+
+    t1 = t1 * dst_reg[0];
+
+    v_if (sfpi::abs(val) > 1){
+        t1 = 1.570796327f - t1;
+    }
+    v_endif;
+
+    v_if(val < 0 ){
+        t1 = -t1;
+    }
+    v_endif;
+
+    return t1;
+}
+
+template <bool APPROXIMATION_MODE, int ITERATIONS = 4>
+inline void calculate_atan()
+{
+    // SFPU microcode
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat val = dst_reg[0];
+        val = sfpu_atan_maclaurin_series<APPROXIMATION_MODE>(val);
+        dst_reg[0] = val;
+        dst_reg++;
+    }
 }
 
 template <bool APPROXIMATION_MODE>
-void cosine_init() {
-    ;
+sfpi_inline vFloat sfpu_asine_maclaurin_series(vFloat val)
+{
+    // input for [-1:1]
+    // Mclauren series
+    // arcsin(x) = x + [(1/2) *x^3/3] + [(1 * 3) / (2 * 4) * x^5 / 5] + [(1 * 3 * 5) / (2 * 4 * 6) * x^7 / 7 ] + ...
+    // arcsin(x) ≈ x + (1/6) * x^3 + (3/40) * x^5 + (5/112) * x^7 + (35/1152) * x^9 + (63/2816) * x^11
+
+    vFloat tmp = val;
+    vFloat val_square = val * val;
+    // x
+    vFloat output = tmp;
+    // (1/6) * x^3
+    tmp = tmp * val_square;
+    output += 0.166666666 * tmp;
+    // (3/40) * x^5
+    tmp = tmp * val_square;
+    output +=  0.075 * tmp;
+
+    //(5/112) * x^7
+    tmp = tmp * val_square;
+    output += 0.044642857 * tmp;
+
+    // (35/1152) *x^9
+    tmp = tmp * val_square;
+    output += 0.03038194 * tmp;
+
+    //(63/2816) * x^11
+    tmp = tmp * val_square;
+    output += 0.02237216 * tmp;
+
+    // Write out output
+    return output;
 }
 
-template <bool APPROXIMATION_MODE>
-void tan_init() {
-    ;
+template <bool APPROXIMATION_MODE, int ITERATIONS = 4>
+inline void calculate_asin()
+{
+    // SFPU microcode
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat v = dst_reg[0];
+        v = sfpu_asine_maclaurin_series<APPROXIMATION_MODE>(v);
+        dst_reg[0] = v;
+        dst_reg++;
+    }
 }
+
+template <bool APPROXIMATION_MODE, int ITERATIONS = 4>
+inline void calculate_acos()
+{
+    // SFPU microcode
+    // acos = (pi/2 - asin)
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat v = dst_reg[0];
+        v = sfpu_asine_maclaurin_series<APPROXIMATION_MODE>(v);
+        v = (1.570796326794) - v;
+        dst_reg[0] = v;
+        dst_reg++;
+    }
+}
+
 }  // namespace sfpu
 }  // namespace ckernel
