+++ tt_metal/hw/ckernels/grayskull/metal/llk_api/llk_sfpu/llk_math_eltwise_unary_sfpu_trigonometry.h
@@ -17,7 +17,7 @@ namespace ckernel {
 //sine
 template <bool APPROXIMATE>
 inline void llk_math_eltwise_unary_sfpu_sine_init() {
-    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>(sfpu::sine_init<APPROXIMATE>);
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
 }
 
 template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
@@ -32,7 +32,7 @@ inline void llk_math_eltwise_unary_sfpu_sine_op(uint dst_index, int vector_mode 
 //cosine
 template <bool APPROXIMATE>
 inline void llk_math_eltwise_unary_sfpu_cosine_init() {
-    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>(sfpu::cosine_init<APPROXIMATE>);
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
 }
 
 template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
@@ -47,7 +47,7 @@ inline void llk_math_eltwise_unary_sfpu_cosine_op(uint dst_index, int vector_mod
 //tangent
 template <bool APPROXIMATE>
 inline void llk_math_eltwise_unary_sfpu_tan_init() {
-    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>(sfpu::tan_init<APPROXIMATE>);
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
 }
 
 template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
@@ -57,6 +57,48 @@ inline void llk_math_eltwise_unary_sfpu_tan_op(uint dst_index, int vector_mode =
                                 ckernel::sfpu::calculate_sfpu_trig<SfpuType::tan, APPROXIMATE>,
                                 dst_index, vector_mode);
 
+}
+
+//asin
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_asin_init() {
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
+}
+
+template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
+inline void llk_math_eltwise_unary_sfpu_asin(uint dst_index, int vector_mode = (int)VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu_0_param<APPROXIMATE, Dst>
+                                (ckernel::sfpu::calculate_asin<APPROXIMATE>,
+                                ckernel::sfpu::calculate_asin<APPROXIMATE>,
+                                dst_index, vector_mode);
+}
+
+//acos
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_acos_init() {
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
+}
+
+template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
+inline void llk_math_eltwise_unary_sfpu_acos(uint dst_index, int vector_mode = (int)VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu_0_param<APPROXIMATE, Dst>
+                                (ckernel::sfpu::calculate_acos<APPROXIMATE>,
+                                ckernel::sfpu::calculate_acos<APPROXIMATE>,
+                                dst_index, vector_mode);
+}
+
+//atan
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_atan_init() {
+    llk_math_eltwise_unary_sfpu_init<APPROXIMATE>();
+}
+
+template <bool APPROXIMATE, DstSync Dst = DstSync::SyncFull>
+inline void llk_math_eltwise_unary_sfpu_atan(uint dst_index, int vector_mode = (int)VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu_0_param<APPROXIMATE, Dst>
+                                (ckernel::sfpu::calculate_atan<APPROXIMATE>,
+                                ckernel::sfpu::calculate_atan<APPROXIMATE>,
+                                dst_index, vector_mode);
 }
 
 }
