+++ tt_metal/include/compute_kernel_api.h
@@ -21,6 +21,7 @@
 #include "llk_math_matmul_api.h"
 #include "llk_math_unary_datacopy_api.h"
 #include "llk_math_binary_api.h"
+#include "llk_math_unary_sfpu_api.h"
 #include "llk_math_eltwise_unary_sfpu_expm1.h"
 #include "llk_math_eltwise_unary_sfpu_rsqrt.h"
 #include "llk_math_eltwise_unary_sfpu_signbit.h"
@@ -29,14 +30,14 @@
 #include "llk_math_eltwise_unary_sfpu_heaviside.h"
 #include "llk_math_eltwise_unary_sfpu_silu.h"
 #include "llk_math_eltwise_unary_sfpu_sigmoid.h"
-#include "llk_math_eltwise_unary_sfpu_tanh.h"
-#include "llk_math_eltwise_unary_sfpu_square.h"
-#include "llk_math_eltwise_unary_sfpu_log.h"
-#include "llk_math_eltwise_unary_sfpu_abs.h"
 #include "llk_math_eltwise_unary_sfpu_sign.h"
-#include "llk_math_eltwise_unary_sfpu_max.h"
 #include "llk_math_eltwise_unary_sfpu_comp.h"
 #include "llk_math_eltwise_unary_sfpu_trigonometry.h"
+// #include "llk_math_eltwise_unary_sfpu_tanh.h"
+// #include "llk_math_eltwise_unary_sfpu_square.h"
+// #include "llk_math_eltwise_unary_sfpu_log.h"
+// #include "llk_math_eltwise_unary_sfpu_abs.h"
+// #include "llk_math_eltwise_unary_sfpu_max.h"
 #include "llk_math_reduce_api.h"
 #define MATH(x) x
 #define MAIN math_main()
